Feature/US_33_Edit_office_&_US_32_Change_office’s_status (#16)
* fix(EmployeeUI): fix authorization flow in offices-service

* feat(OfficesAPI): implement [HttpPut] UpdateAsync endpoint, project upd

Add Update(), OfficeExistsAsync() and OfficeAddressExistsAsync() methods to the OfficesRepository and IOfficesRepository
Add UpdateAsync() method to the OfficeService and IOfficeService
Add global AaNoTracking option for the EFCore
Update the Office model validation flow

* refactor(OfficesAPI): update logging and the Office model validation

Update logging in the OfficesAPI
Update the Office model validation in the UpdateAsync method in the OfficesController

* feat(EmployeeUI): implement an office info editing

Add edit-office-form component
Add updateOffice() method to the the office.service
Update the office-card component to display the edit-office form
Update the create-office form visuals

* feat(EmployeeUI): update the office info modal window behavior

The office info modal window tracks changes to office information after entering editing mode via the “Edit” button. A change is only counted when the user clicks the “Save changes” button in editing mode, and the offices table is updated accordingly.
Additionally, while in editing mode, this window cannot be closed by clicking the backdrop, the close (×) icon, or pressing the Esc key.

* refactor(EmployeeUI): update office forms visuals

Update the create-office form visuals
Update the edit-office form visuals
Update the office-card visuals

* feat(): implement doctors profiles deactivation flow

Update IdentityServer configuration in the AuthorizationAPI
Update JWT token bearer settings in the ProfilesAPI
Add deactivateDoctorsByOfficeId() method to the doctor.service in the EmployeeUI
Implement [HttpPut("deactivate/byOfficeId/{officeId:Guid}")] DeactivateDoctorsByOfficeIdAsync endpoint in the DoctorsController
Add GetAllByOfficeIdAsync() method to the DoctorsRepository and IDoctorsRepository
Add GetAllByOfficeIdAsync() method to the ReceptionistRepository and IReceptionistRepository
Add DeactivateProfilesByOfficeIdAsync() method to the DoctorService and IDoctorService

* refactor(): add docu. for the OfficeController and minor visuals update

Update validation errors visuals for the registryPhoneNumber field of the create-office form in the EmployeeUI
Update validation errors visuals in the edit-office form in the EmployeeUI
Add documentation for the OfficesController in the OfficeAPI