Feature/CI_CD_Setup (#19)
*Add CI workflow
CI runs automatically on PRs to the main or develop branches. It can also be executed manually through the Actions tab in the repository.
Contains the following jobs:
 - Build all projects (including APIs and SPAs)
 - Run tests for APIs (excluding integration tests)
 - Show test results

*Add CD workflow
CD runs automatically on push to the main branch. It can also be executed manually through the Actions tab in the repository.
Contains the following steps:
 - Build images
 - Push images to DockerHub
 - Update docker-compose.yml file on the remote server
 - Stop running containers on the remote server
 - Delete old project images on the remote server
 - Pull new images to the remote server
 - Launch new containers on the remote server

*Update tests
Add the Category attribute to all test classes. This helps to choose, which tests should be run in the CI workflow.

*Update dcoker-compose.yml file
Add links to projects images in DockerHub repository.