Feature/Appointments_API_boilerplate (#26)
feat(): Create Appointments API boilerplate (#26)

Caution

Do not use these API endpoints until they are properly implemented.
Implementation status can be checked in the project wiki.

Update NuGet packages in solution

Add Appointments API boilerplate

This boilerplate represents the future structure of the API and is not intended for use at this stage.
The only reason it was created is to update the existing CI/CD pipeline and prepare for the implementation of the gateway API.